Not to depend on BCC python binding in the code generator
