+++ CMakeLists.txt
@@ -47,10 +47,9 @@ ADD_CUSTOM_COMMAND(
 )
 LIST(APPEND DTRACE_FILES "${CMAKE_BINARY_DIR}/h2o-probes.d")
 
-SET(H2O_BINARY "${CMAKE_SOURCE_DIR}/../../h2o/h2o/build/h2o") # FIXME
 ADD_CUSTOM_COMMAND(
     OUTPUT  "${CMAKE_BINARY_DIR}/quic.cc"
-    COMMAND "${CMAKE_SOURCE_DIR}/gen-bpf.py" "${H2O_BINARY}" "${CMAKE_BINARY_DIR}" "${CMAKE_BINARY_DIR}/quic.cc"
+    COMMAND "${CMAKE_SOURCE_DIR}/gen-bpf.py" "${CMAKE_BINARY_DIR}" "${CMAKE_BINARY_DIR}/quic.cc"
     DEPENDS ${DTRACE_FILES}
     DEPENDS "${CMAKE_SOURCE_DIR}/gen-bpf.py"
 )
